backend/requirements.txt
# Core dependencies
chromadb>=0.4.0
chonkie>=0.1.0
fastapi>=0.104.0
uvicorn>=0.24.0
ollama>=0.2.0
transformers>=4.35.0
torch>=2.0.0
numpy>=1.24.0

# File parsing
PyPDF2>=3.0.0
python-docx>=0.8.11
Pillow>=10.0.0
easyocr>=1.7.0

# API dependencies
python-multipart>=0.0.6
pydantic>=2.0.0

# Development dependencies
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
isort>=5.12.0
flake8>=6.0.0